// frontend-v2/src/app/core/api/models/categoria-input.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CategoriaDisponibilidadeDto } from '../models/categoria-disponibilidade-dto';

/**
 * Dados da categoria
 */
export interface CategoriaInput {

  /**
   * Descricao da categoria
   */
  descricao?: string;

  /**
   * Objeto contendo os dias da semana em que a categoria está disponivel
   */
  disponivel?: CategoriaDisponibilidadeDto;

  /**
   * Horario final (hh:mm) em que a categoria fica disponivel
   */
  fim?: string;

  /**
   * ID da culinaria na qual pertence a categoria
   */
  id_culinaria: number;

  /**
   * Horario inicial (hh:mm) em que a categoria fica disponivel
   */
  inicio?: string;

  /**
   * Nome da categoria
   */
  nome: string;

  /**
   * Usada em pizzas: '', 'M' (valor medio) ou 'V' (maior valor)
   */
  opcao_meia?: string;

  /**
   * Lista de nomes das opcoes da categoria
   */
  opcoes: Array<string>;

  /**
   * Ordem da categoria no cardapio (unica por organizacao)
   */
  ordem?: number;
}