// frontend-v2/src/app/core/api/models/culinaria-output.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Dados da culinaria
 */
export interface CulinariaOutput {

  /**
   * ID da culinaria
   */
  id?: number;

  /**
   * Indica se permite meio a meio
   */
  meioMeio?: boolean;

  /**
   * Nome da culinaria
   */
  nome?: string;
}